dashboard_service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
